ai-router/requirements.txt
sentence-transformers
scikit-learn
gradio
pandas
requests
numpy
python-dotenv
